+++ interface/catalog/creodias/scenes.go
@@ -126,7 +126,7 @@ func (s *Provider) SearchScenes(ctx context.Context, area *entities.AreaToIngest
 		// Create scene
 		scenes[i] = &entities.Scene{
 			Scene: common.Scene{
-				SourceID: rawscene.Properties.Identifier[:len(rawscene.Properties.Identifier)-len(".SAFE")],
+				SourceID: strings.TrimSuffix(rawscene.Properties.Identifier, ".SAFE"),
 				Data: common.SceneAttrs{
 					UUID:         rawscene.Uuid,
 					Date:         date,